// packages/destination-actions/src/destinations/calliper/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Settings {
  /**
   * Your company id in Calliper. During the onboarding you can find it by selecting Segment integration. After onboarding is complete you can find it by opening "Data Sources" in the main menu, pressing "New Data Source" and pressing "Connect" on a Segment card.
   */
  companyId: string
  /**
   * Your Segment Key in Calliper. Please follow the instructions to find the Company id, secret key is located in a field next to it.
   */
  segmentKey: string
}